requirements.txt
gunicorn
psycopg2-binary==2.9.9
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
Werkzeug==2.3.7
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
Pillow==11.3.0
filetype==1.2.0
python-magic==0.4.27
waitress==3.0.2
Jinja2==3.1.6
SQLAlchemy==2.0.41
greenlet==3.2.3
itsdangerous==2.2.0
click==8.2.1
blinker==1.9.0
MarkupSafe==3.0.2
charset-normalizer==3.4.2
idna==3.10
urllib3==2.5.0
certifi==2025.7.14
soupsieve==2.7
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib